requirements.txt
requests==2.31.0
pandas==2.1.4
pydantic==2.5.2
python-dotenv==1.0.0
flask==3.0.0
openpyxl==3.1.2
gunicorn==21.2.0
